rproxy: tcp: add subcommand support
